Fix Royal seed pod teleport false positive
When teleporting using the seed pod, the player's position is first
changed (but `INSIDE_WILDERNESS` and `MULTIWAY_COMBAT` are not), which
caused alerting to happen on the second tick after arriving on the
teleport tile. This change adds that tile to `FORCE_MOVE_IGNORE_AREA` so
it doesn't fire false positive alerts.