Remove area mismatch alerting
Thus far it has only yielded false positives, and continues to cause
unreproduceable reports. While this feature might prove useful in the
future, it would require diagnosing and fixing the cause of these
errors. For now, I'll keep the rest of the code associated with these
alerts, but remove the chat message in the event that the reporting
command might prove useful in the meantime.